# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: AI CRM & Sales Automation API
  version: "1.0.0-draft"
  description: >
    Draft contract for Phase 08. Every operation carries an `x-roles` extension
    (see README §2) documenting which OrgRole values may call it — enforcement
    happens in the Rbac Guard (architecture/README.md §6.1), not here.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      security: []
      x-roles: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      security: []
      x-roles: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      # Fixed during M1 implementation: README §2 already documented this as
      # "any authenticated (refresh token)" — i.e. authenticated BY the
      # refresh token itself, not by a still-valid Bearer access token (which
      # would be expired in the exact case a client calls this). The original
      # draft inherited the global bearerAuth requirement by omission; that
      # was a contract bug, not a real requirement, so `security: []` +
      # `x-roles: []` match the README's intent instead of the literal draft.
      security: []
      x-roles: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/password-reset/request:
    post:
      tags: [Auth]
      operationId: requestPasswordReset
      security: []
      x-roles: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202": { description: Reset email sent if the account exists }
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      operationId: confirmPasswordReset
      security: []
      x-roles: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, format: password, minLength: 8 }
      responses:
        "204": { description: Password updated }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/email/verify:
    post:
      tags: [Auth]
      operationId: verifyEmail
      x-roles: [OWNER, ADMIN, SALES_MANAGER, SALES_REP, VIEWER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "204": { description: Email verified }

  # ---------------------------------------------------------------------
  # Organizations & Members — FR-006 to FR-012
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        emailVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
...